pyproject.toml
[tool.poetry]
name = "golden-wind"
version = "0.1.0"
description = ""
authors = ["Sergio Fuentes de Uña <serede.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-telegram-bot = "^21.10"
pymupdf = "^1.25.3"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
golden-wind = "golden_wind:main"
